requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
firebase-admin
python-dotenv
flask_cors